toolkit/components/passwordmgr/test/mochitest/test_autocomplete_password_open.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test password field autocomplete footer with and without logins</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="../../../satchel/test/satchel_common.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<script>
var chromeScript = runChecksAfterCommonInit();

let origin = window.location.origin;
addLoginsInParent(
  [origin, "", null, "tempuser1", "temppass1", "uname", "pword"],
  [origin, "", null, "tempuser2", "temppass2", "uname", "pword"]);
</script>
<p id="display"></p>

<div id="content">
  <h1>Sign in</h1>
  <input id="uname"/>
  <input id="pword" type="password"/>
</div>

<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Login Manager: Test password field autocomplete footer with and without logins **/

let uname = document.getElementById("uname");
let pword = document.getElementById("pword");

add_task(async function test_no_autofill() {
  // Make sure initial form is empty as autofill shouldn't happen in the sandboxed frame.
  checkLoginForm(uname, "", pword, "");
  let popupState = await getPopupState();
  is(popupState.open, false, "Check popup is initially closed");
});

add_task(async function test_two_logins() {
  let shownPromise = promiseACShown();
  uname.focus();
  await shownPromise;

  shownPromise = promiseACShown();
  synthesizeKey("KEY_Tab"); // popup on the password field should open upon focus
  let results = await shownPromise;

  let popupState = await getPopupState();
  is(popupState.selectedIndex, -1, "Check no entries are selected upon opening");

  let expectedMenuItems = [
    "tempuser1",
    "tempuser2",
  ];
  checkAutoCompleteResults(results, expectedMenuItems, window.location.host, "Check all menuitems are displayed correctly.");

  checkLoginForm(uname, "", pword, "");
  let removedPromise = promiseStorageChanged(["removeAllLogins"]);
  await LoginManager.removeAllUserFacingLogins();
  await removedPromise;
});

add_task(async function test_zero_logins() {
  uname.focus();

  let shownPromise = promiseACShown().then(() => ok(false, "Should not have shown"));
  // Popup on the password field should NOT automatically open upon focus when there are no saved logins.
  synthesizeKey("KEY_Tab"); // focus the password field
  SimpleTest.requestFlakyTimeout("Giving a chance for the unexpected popup to show");
  let autocompleteItems = await Promise.race([
    shownPromise,
    new Promise(resolve => setTimeout(resolve, 2000)), // Wait 2s for the popup to appear
  ]);

  let popupState = await getPopupState();
  is(popupState.open, false, "Check popup is still closed");

  checkLoginForm(uname, "", pword, "");
  info("arrow down should still open the popup");
  autocompleteItems = await openAutocompletePopup();
  checkAutoCompleteResults(autocompleteItems, [], window.location.host, "Check only footer is displayed.");
  checkLoginForm(uname, "", pword, "");
});
</script>
</pre>
</body>
</html>
